parserelle-iot-l2l3-actionneur/requirements.txt
cryptography==3.3.2
PyJWT==1.6.4
pyOpenSSL==18.0.0
rfc3339==6.0
paho-mqtt==1.3.1
pyserial==3.4
